Recognizer centrally updates input, new pool lists
  - Rather than having each 'thread' update its
    own input pointer, they all share a 'global'
    pointer.

  - Simpler list implementation for managing
    'thread' pools. This significantly reduces
    instructions executed when managing pools.

+++ tests/test_slist/wrapper_funcs.h
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-void * __real_xmalloc(unsigned int sz);
+void * __real_xmalloc(size_t sz);
 void __real_free(void *);
-void * __wrap_xmalloc(unsigned int sz);
+void * __wrap_xmalloc(size_t sz);
 void __wrap_free(void *);
